refactor: StatusView 위치 이동 및 PrezelLottie 컴포넌트 분리
* refactor: StatusView 패키지 위치 변경 및 코드 정리
    * `StatusView`를 `com.team.prezel.core.ui`에서 `com.team.prezel.core.ui.component` 패키지로 이동하였습니다.
    * `StatusView.kt` 내부에 정의되어 있던 `StatusLottie`와 로딩 프리뷰 코드를 제거하였습니다.

* feat: 공용 Lottie 컴포저블 `PrezelLottie` 구현
    * Lottie 애니메이션을 일관되게 사용하기 위해 `PrezelLottie` 컴포넌트를 새롭게 추가하였습니다.
    * `isPlaying`, `iterations`, `speed`, `onAnimationFrame` 등 다양한 속성을 제어할 수 있도록 파라미터를 제공합니다.

* refactor: HistoryEmptyContent 내 StatusView 참조 수정
    * `StatusView` 패키지 위치 변경에 따른 import 문을 업데이트하였습니다.
